Correção nos métodos async de updateRating/Consumer

+++ service_quartos/src/main/java/com/sistemareseva/service_quartos/client/broker/dto/UpdateRatingRequest.java
@@ -1,8 +1,10 @@
 package com.sistemareseva.service_quartos.client.broker.dto;
 
+import java.math.BigDecimal;
+
 public record UpdateRatingRequest(
     Long idQuarto,
-    Double rating
+    BigDecimal rating
 ) {
     
 }
